Optimization 2
We realized several numbers do not deserve so many for-loop iterations. Even numbers after 2 do not need any for loop iteration since they are not prime, and neither do odd numbers for which the divisibility counter exceeds 2. We added the necessary logic and short-circuiting to save time for these scenarios.